feat : 3주차

+++ book-order/src/main/java/com/example/bookorder/domain/Money.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 public class Money {
     private BigDecimal value;
 
-    public Money(BigDecimal value) {
-        this.value = value;
+    public Money(int value) {
+        this.value = BigDecimal.valueOf(value);
     }
 }
